Created the Kick-Stick admin tool. Updated guns to have bullet spread. Moved base-logic for bullets to BaseBullet, with an abstract onTravel(loc) method to perform actions per-location. Moved UserManager interface to IUserManager Made UserManager generic class, which now (in conjunction with the new generic minigame class, and gameconnectionlistener) handles player data loading and removing. Guns now reload when you attempt to drop them (by default) Added SneakyThrow from Spigot Fixed the issue with cooldowns counting backwards

+++ src/main/java/com/caved_in/commons/entity/Entities.java
@@ -458,7 +458,7 @@ public class Entities {
 	 * @param radius radius to scan around the location.
 	 * @return a set of living entities which were near the location
 	 */
-	public static Set<LivingEntity> getLivingEntitiesNearLocation(Location center, int radius) {
+	public static Set<LivingEntity> getLivingEntitiesNearLocation(Location center, double radius) {
 		Set<LivingEntity> entities = new HashSet<>();
 		for (LivingEntity entity : center.getWorld().getLivingEntities()) {
 			if (Locations.isEntityInRadius(center, radius, entity)) {
